refactor stateclient - change name to stategetter
